+++ js/map.js
@@ -1,4 +1,5 @@
 import {unlockForm} from './form-initialization.js';
+// import { compareOffers, getOfferRank } from './map-filter.js';
 
 const SIMILAR_OFFER_COUNT = 10;
 const LAT_TOKIO = 35.68950;
@@ -68,122 +69,17 @@ const initMap = () => {
   resetMap();
 };
 
-const getOfferRank = (offer) => {
-  const housingType = document.querySelector('#housing-type');
-  const housingPrice = document.querySelector('#housing-price');
-  const housingRooms = document.querySelector('#housing-rooms');
-  const housingGuests = document.querySelector('#housing-guests');
-  const housingFeaturesWifi = document.querySelector('#filter-wifi');
-  const housingFeaturesDishwasher = document.querySelector('#filter-dishwasher');
-  const housingFeaturesParking = document.querySelector('#filter-parking');
-  const housingFeaturesWasher = document.querySelector('#filter-washer');
-  const housingFeaturesElevator = document.querySelector('#filter-elevator');
-  const housingFeaturesConditioner = document.querySelector('#filter-conditioner');
+// const typeFilter = (offer) => {
+//   if (offer.offer.type === document.querySelector('#housing-type').value) {
+//     return true;
+//   }
+// };
 
-  let rank = 0;
-
-  if (offer.offer.type === housingType.value) {
-    rank += 10;
-  }
-  if (offer.offer.rooms === +housingRooms.value) {
-    rank += 1;
-  }
-  if (offer.offer.guests === +housingGuests.value) {
-    rank += 1;
-  }
-  if (housingPrice.value === 'low' && offer.offer.price<10000) {
-    rank +=1;
-  }
-  if (housingPrice.value === 'middle' && offer.offer.price>=10000 && offer.offer.price <= 50000) {
-    rank +=1;
-  }
-  if (housingPrice.value === 'high' && offer.offer.price>50000) {
-    rank +=1;
-  }
-  if (typeof offer.offer.features === 'object') {
-    if (housingFeaturesWifi.checked && offer.offer.features.includes(housingFeaturesWifi.value)) {
-      rank +=1;
-    }
-    if (housingFeaturesDishwasher.checked && offer.offer.features.includes(housingFeaturesDishwasher.value)) {
-      rank +=1;
-    }
-    if (housingFeaturesParking.checked && offer.offer.features.includes(housingFeaturesParking.value)) {
-      rank +=1;
-    }
-    if (housingFeaturesWasher.checked && offer.offer.features.includes(housingFeaturesWasher.value)) {
-      rank +=1;
-    }
-    if (housingFeaturesElevator.checked && offer.offer.features.includes(housingFeaturesElevator.value)) {
-      rank +=1;
-    }
-    if (housingFeaturesConditioner.checked && offer.offer.features.includes(housingFeaturesConditioner.value)) {
-      rank +=1;
-    }
-  }
-  return rank;
-};
-
-const compareOffers = (offerA, offerB) => {
-  const rankA = getOfferRank(offerA);
-  const rankB = getOfferRank(offerB);
-
-  return rankB - rankA;
-};
-
-const onChangeType = (cb) => {
-  document.querySelector('#housing-type').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-};
-
-const onChangePrice = (cb) => {
-  document.querySelector('#housing-price').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-};
-
-const onChangeRoom = (cb) => {
-  document.querySelector('#housing-rooms').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-};
-
-const onChangeGuests = (cb) => {
-  document.querySelector('#housing-guests').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-};
-
-const onChangeFeature = (cb) => {
-  document.querySelector('#filter-wifi').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-  document.querySelector('#filter-dishwasher').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-  document.querySelector('#filter-parking').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-  document.querySelector('#filter-washer').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-  document.querySelector('#filter-elevator').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-  document.querySelector('#filter-conditioner').addEventListener('change', () => {
-    markerGroup.clearLayers();
-    cb();
-  });
-};
+// const typeRoomsFilter = (offer) => {
+//   if((offer.offer.type === document.querySelector('#housing-type').value) && (offer.offer.rooms === +document.querySelector('#housing-rooms').value)) {
+//     return true;
+//   }
+// };
 
 const insertOffers = (array, renderOffer) => {
 
@@ -193,24 +89,72 @@ const insertOffers = (array, renderOffer) => {
     iconAnchor: [20, 40],
   });
 
-  array
-    .slice()
-    .sort(compareOffers)
-    .slice(0,SIMILAR_OFFER_COUNT)
-    .forEach((element) => {
-      const regularPinMarker = L.marker(
-        {
-          lat: element.location.lat,
-          lng: element.location.lng,
-        },
-        {
-          icon: regularPinIcon,
-        }
-      );
-      regularPinMarker
-        .addTo(markerGroup)
-        .bindPopup(renderOffer(element));
-    });
+  let targetArray = array
+    .filter((element) =>{
+      if (element.offer.type === document.querySelector('#housing-type').value) {
+        return true;
+      } else
+      if(element.offer.rooms === +document.querySelector('#housing-rooms').value) {
+        return true;
+      }
+      if (document.querySelector('#housing-price').value === 'low' && element.offer.price<10000) {
+        return true;
+      } else
+      if (document.querySelector('#housing-price').value === 'middle' && element.offer.price>=10000 && element.offer.price <= 50000) {
+        return true;
+      } else
+      if (document.querySelector('#housing-price').value === 'high' && element.offer.price>50000) {
+        return true;
+      } else
+      if (element.offer.guests === +document.querySelector('#housing-guests').value) {
+        return true;
+      }
+      // if (typeof element.offer.features === 'object') {
+      //   if (document.querySelector('#filter-wifi').checked && element.offer.features.includes(document.querySelector('#filter-wifi').value)) {
+      //     return true;
+      //   }
+      // if (housingFeaturesDishwasher.checked && offer.offer.features.includes(housingFeaturesDishwasher.value)) {
+      //   rank +=1;
+      // }
+      // if (housingFeaturesParking.checked && offer.offer.features.includes(housingFeaturesParking.value)) {
+      //   rank +=1;
+      // }
+      // if (housingFeaturesWasher.checked && offer.offer.features.includes(housingFeaturesWasher.value)) {
+      //   rank +=1;
+      // }
+      // if (housingFeaturesElevator.checked && offer.offer.features.includes(housingFeaturesElevator.value)) {
+      //   rank +=1;
+      // }
+      // if (housingFeaturesConditioner.checked && offer.offer.features.includes(housingFeaturesConditioner.value)) {
+      //   rank +=1;
+      // }
+      // }
+    }
+    );
+
+  if (targetArray.length > SIMILAR_OFFER_COUNT) {
+    targetArray = targetArray.slice(0, SIMILAR_OFFER_COUNT);
+  }
+
+  targetArray.forEach((element) => {
+    const regularPinMarker = L.marker(
+      {
+        lat: element.location.lat,
+        lng: element.location.lng,
+      },
+      {
+        icon: regularPinIcon,
+      }
+    );
+    regularPinMarker
+      .addTo(markerGroup)
+      .bindPopup(renderOffer(element));
+  });
+
+
+};
+const clearLayers = () => {
+  markerGroup.clearLayers();
 };
 
-export {initMap, insertOffers, resetMap, onChangeType, onChangeRoom, onChangePrice, onChangeGuests, onChangeFeature};
+export {initMap, insertOffers, resetMap, clearLayers};
